requirements.txt
Django
django-cors-headers
djangorestframework
gunicorn
python-decouple
sqlparse
psycopg2-binary
redis
django-redis
celery
openpyxl